--- a/libvirt/tests/cfg/migration/migration_performance_tuning/migration_zerocopy.cfg
+++ b/libvirt/tests/cfg/migration/migration_performance_tuning/migration_zerocopy.cfg
@@ -32,7 +32,6 @@
     check_str_local_log = '"capability":"zero-copy-send","state":true'
     func_supported_since_libvirt_ver = (8, 0, 0)
     action_during_mig = '[{"func": "check_qemu_mem_lock_hard_limit", "after_event": "iteration: '1'", "func_param": 'params'}, {"func": "set_migrate_speed_to_high", "func_param": "params"}]'
-    migrate_speed_high = '8796093022207'
     migrate_speed = "5"
     stress_package = "stress"
     stress_args = "--cpu 8 --io 8 --vm 4 --vm-bytes 128M --timeout 20s"
@@ -48,7 +47,7 @@
     variants test_case:
         - with_memtune:
             hard_limit = 10485760
-            vm_attrs = {'memtune': {'hard_limit': ${hard_limit}, 'hard_limit_unit': 'KiB'}}
+            memtune_options = "--hard-limit ${hard_limit}"
         - without_memtune:
     variants:
         - zerocopy_and_parallel_and_tcp_transport: