--- a/libvirt/tests/cfg/migration/migration_performance_tuning/migration_zerocopy_abort.cfg
+++ b/libvirt/tests/cfg/migration/migration_performance_tuning/migration_zerocopy_abort.cfg
@@ -24,7 +24,7 @@
     migrate_desturi_type = "tcp"
     virsh_migrate_desturi = "qemu+tcp://${migrate_dest_host}/system"
     func_supported_since_libvirt_ver = (8, 0, 0)
-    action_during_mig = '[{"func": "check_qemu_mem_lock_hard_limit", "after_event": "iteration: '1'", "func_param": 'params'}, {"func": "virsh.domjobabort", "func_param": "'%s' % params.get('migrate_main_vm')"}]'
+    action_during_mig = '[{"func": "check_qemu_mem_lock_hard_limit", "after_event": "iteration: '1'", "func_param": 'params'}, {"func": "virsh.domjobabort", "func_param": "'%s' % params.get('migrate_main_vm')"}, {"func": "set_migrate_speed_to_high", "func_param": "params"}]'
     virsh_migrate_extra = "--zerocopy --parallel"
     check_hard_limit = "yes"
     status_error = "yes"
@@ -45,5 +45,5 @@
     variants test_case:
         - with_memtune:
             hard_limit = 10485760
-            vm_attrs = {'memtune': {'hard_limit': ${hard_limit}, 'hard_limit_unit': 'KiB'}}
+            memtune_options = "--hard-limit ${hard_limit}"
         - without_memtune: