Update to use virsh.memtune_set() to set memtune hard limit.

Signed-off-by: cliping <lcheng@redhat.com>